requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-appconf==1.0.3
django-cors-headers==2.5.2
django-filter==2.1.0
django-oauth-toolkit==1.2.0
djangorestframework==3.9.2
geographiclib==1.49
geopy==1.19.0
idna==2.8
importlib-metadata==0.23
kombu==4.6.6
more-itertools==7.2.0
oauthlib==3.0.1
#psycopg2==2.8.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
validate-email==1.3
vine==1.3.0
zipp==0.6.0
